requirements.txt
feedparser
tweepy
python-telegram-bot
groq
google-generativeai
pyyaml
python-dotenv
